Add Data Entries 90% Finished
- Able to add data entries to all tables except relation related tables (missing "PlayerData")
- Updates Relation Tables as well.

To do:
Add Player Data and hook it up to its respective relation tables.

Possible bug:
The current implementation for the primary keys of the relation tables may bug out the code later on. Composite keys aren't a thing, so I settled for just making a general "RelationKey" for all of them. This should not affect anything right now, but if it does, let me know.

Once again, you can check the database file using this website:
https://inloop.github.io/sqlite-viewer/